+++ src/app/housing-location/housing-location.component.ts
@@ -1,4 +1,6 @@
-import { Component } from '@angular/core';
+import {Component, Input} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {HousingLocation} from '../housinglocation';
 
 @Component({
   selector: 'app-housing-location',
@@ -10,6 +12,7 @@ import { Component } from '@angular/core';
   `,
   styleUrls: ['./housing-location.component.css'],
 })
-export class HousingLocationComponent {
 
+export class HousingLocationComponent {
+  @Input() housingLocation!: HousingLocation;
 }